Добавляет abstract-smart-component.js, наводит порядок в директории event-edit
