+++ src/components/abstract-smart-component.js
@@ -1,0 +1,20 @@
+import AbstractComponent from "./abstract-component";
+
+export default class AbstractSmartComponent extends AbstractComponent {
+  recoverListeners() {
+    throw new Error(`Abstract method not implemented: recoveryListeners`);
+  }
+
+  rerender() {
+    const oldElement = this.getElement();
+    const parent = oldElement.parentElement;
+
+    this.removeElement();
+
+    const newElement = this.getElement();
+
+    parent.replaceChild(newElement, oldElement);
+
+    this.recoverListeners();
+  }
+}
