--- src/components/page-main/events/event-edit/event-photo.js
@@ -1,7 +1,0 @@
-const createEventPhotoTemplate = (src) => {
-  return (
-    `<img class="event__photo" src="${src}" alt="Event photo">`
-  );
-};
-
-export {createEventPhotoTemplate};
